Add command line. Re-structure.
PyPA say: don't use your package as a namespace. So, I'm going to use
`textbox.storage` as a namespace, instead.

I'm trying to avoid dependencies, but the command line is going to be
the UI and API for quite a while and I lack the time to deliver quality
using raw argparse.